fix code check

+++ pc/services/hap/entity/include/hw_block_data.h
@@ -22,16 +22,16 @@ namespace OHOS {
 namespace SignatureTools {
 class HwBlockData {
 public:
-    HwBlockData(int32_t blockNum, int32_t blockStart);
+    HwBlockData(int32_t blockNum, int64_t blockStart);
 
     int32_t GetBlockNum();
     void SetBlockNum(int32_t blockNum);
-    int32_t GetBlockStart();
-    void SetBlockStart(int32_t blockStart);
+    int64_t GetBlockStart();
+    void SetBlockStart(int64_t blockStart);
 
 private:
     int32_t blockNum;
-    int32_t blockStart;
+    int64_t blockStart;
 };
 } // namespace SignatureTools
 } // namespace OHOS
